Decode poll options even when using quill composer
When using quill composer, the raw data provided to the serializer looks like this for the settings: `" title=\\\"This is the poll title\\\" maxvotes=\\\"2\\\""` 
Because the `settingRegex` is not covering this markup, the options fail to be parsed so that the default options are used. I've updated here the settings regex and replace the escaped backslashes in the raw data.